+++ tests/PathTest.php
@@ -8,11 +8,17 @@ use Innmind\Url\{
     AbsolutePath,
     RelativePath,
 };
-use Innmind\BlackBox\PHPUnit\Framework\TestCase;
+use Fixtures\Innmind\Url\Path as FPath;
+use Innmind\BlackBox\{
+    PHPUnit\BlackBox,
+    PHPUnit\Framework\TestCase,
+};
 use PHPUnit\Framework\Attributes\DataProvider;
 
 class PathTest extends TestCase
 {
+    use BlackBox;
+
     public function testInterface()
     {
         $p = Path::of('/foo/bar/');
@@ -79,6 +85,38 @@ class PathTest extends TestCase
         $this->assertInstanceOf(RelativePath::class, Path::of('../somewhere'));
     }
 
+    public function testAbsolutePathsAlwaysStartWithASlash(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(FPath::any())
+            ->prove(function($path) {
+                if ($path->absolute()) {
+                    $this->assertStringStartsWith('/', $path->toString());
+                } else {
+                    $this->assertStringStartsNotWith('/', $path->toString());
+                }
+            });
+    }
+
+    public function testResolveAbsolutility(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(FPath::any(), FPath::any())
+            ->prove(function($a, $b) {
+                $this->assertSame(
+                    $a->absolute() || $b->absolute(),
+                    $a->resolve($b)->absolute(),
+                );
+            });
+    }
+
+    public function testPathStartingWithBackslashIsInvalid()
+    {
+        $this->assert()->throws(
+            static fn() => Path::of('\\\\'),
+        );
+    }
+
     public static function resolutions(): array
     {
         return [
@@ -106,6 +144,11 @@ class PathTest extends TestCase
                 'some/some/target',
                 'some/source',
                 'some/target',
+            ],
+            'relative to the root' => [
+                '/target',
+                '/source',
+                'target',
             ],
         ];
     }